Fix namespaces not allowed to be empty
- Empty namespaces should behave as null/undefined.
- Refactored JsonElementExtensions.
- Tweaked schema error messages.
- Remove code related to enum default.
