+++ tests/AvroSourceGenerator.Tests/EnumTests.cs
@@ -51,7 +51,7 @@ public class EnumTests
     """);
 
     [Theory]
-    [InlineData("null"), InlineData("\"Schema1.Enum.Namespace\""), InlineData("\"schema2.enum.namespace\"")]
+    [InlineData("null"), InlineData("\"\""), InlineData("\"Schema1.Enum.Namespace\""), InlineData("\"schema2.enum.namespace\"")]
     public Task Verify_Namespace(string @namespace) => TestHelper.VerifySourceCode($$"""
     {
         "type": "enum",
@@ -62,7 +62,7 @@ public class EnumTests
     """);
 
     [Theory]
-    [InlineData("\"\""), InlineData("[]")]
+    [InlineData("[]")]
     public Task Verify_Namespace_Diagnostic(string @namespace) => TestHelper.VerifyDiagnostic($$"""
     {
         "type": "enum",
