+++ tests/AvroSourceGenerator.Tests/FixedTests.cs
@@ -78,7 +78,7 @@ public sealed class FixedTests
     """);
 
     [Theory]
-    [InlineData("null"), InlineData("\"Schema1.Throw.Namespace\""), InlineData("\"schema2.throw.namespace\"")]
+    [InlineData("null"), InlineData("\"\""), InlineData("\"Schema1.Throw.Namespace\""), InlineData("\"schema2.throw.namespace\"")]
     public Task Verify_Namespace(string @namespace) => TestHelper.VerifySourceCode($$"""
     {
         "type": "fixed",
@@ -89,7 +89,7 @@ public sealed class FixedTests
     """);
 
     [Theory]
-    [InlineData("\"\""), InlineData("[]")]
+    [InlineData("[]")]
     public Task Verify_Namespace_Diagnostic(string @namespace) => TestHelper.VerifyDiagnostic($$"""
     {
         "type": "fixed",
