+++ tests/AvroSourceGenerator.Tests/RecordTests.cs
@@ -78,7 +78,7 @@ public sealed class RecordTests
     """);
 
     [Theory]
-    [InlineData("null"), InlineData("\"Schema1.Throw.Namespace\""), InlineData("\"schema2.throw.namespace\"")]
+    [InlineData("null"), InlineData("\"\""), InlineData("\"Schema1.Throw.Namespace\""), InlineData("\"schema2.throw.namespace\"")]
     public Task Verify_Namespace(string @namespace) => TestHelper.VerifySourceCode($$"""
     {
         "type": "record",
@@ -89,7 +89,7 @@ public sealed class RecordTests
     """);
 
     [Theory]
-    [InlineData("\"\""), InlineData("[]")]
+    [InlineData("[]")]
     public Task Verify_Namespace_Diagnostic(string @namespace) => TestHelper.VerifyDiagnostic($$"""
     {
         "type": "record",
