+++ tests/AvroSourceGenerator.Tests/Snapshots/EnumTests/EnumTests.Verify_Namespace_namespace=--.verified.txt
@@ -1,0 +1,6 @@
+﻿// <auto-generated/>
+#pragma warning disable CS8618, CS8633, CS8714, CS8775, CS8981
+    [global::System.CodeDom.Compiler.GeneratedCode("AvroSourceGenerator", "1.0.0.0")]
+    public enum TestEnum
+    {
+    }
