+++ tests/AvroSourceGenerator.Tests/Snapshots/EnumTests/EnumTests.Verify_Symbols_Diagnostic_symbols=null.verified.txt
@@ -1,7 +1,7 @@
 ﻿{
   Location: schema.avsc: (0,0)-(5,1),
   Message:
-The schema defined in the JSON is invalid: 'symbols' property is required in schema: {
+The schema defined in the JSON is invalid: 'symbols' property must be an array (found '') in schema: {
     "type": "enum",
     "name": "TestEnum",
     "namespace": "SchemaNamespace",
