+++ tests/AvroSourceGenerator.Tests/Snapshots/ErrorTests/ErrorTests.Verify_Fields_Diagnostic_fields=null.verified.txt
@@ -1,7 +1,7 @@
 ﻿{
   Location: schema.avsc: (0,0)-(5,1),
   Message:
-The schema defined in the JSON is invalid: 'fields' property is required in schema: {
+The schema defined in the JSON is invalid: 'fields' property must be an array (found '') in schema: {
     "type": "error",
     "namespace": "SchemaNamespace",
     "name": "Error",
