+++ tests/AvroSourceGenerator.Tests/Snapshots/FixedTests/FixedTests.Verify_Aliases_Diagnostic_aliases={}.verified.txt
@@ -1,7 +1,7 @@
 ﻿{
   Location: schema.avsc: (0,0)-(6,1),
   Message:
-The schema defined in the JSON is invalid: 'aliases' property must be an array in schema: {
+The schema defined in the JSON is invalid: 'aliases' property must be an array (found '{}') in schema: {
     "type": "fixed",
     "name": "Fixed",
     "namespace": "SchemaNamespace",
