+++ tests/AvroSourceGenerator.Tests/Snapshots/FixedTests/FixedTests.Verify_Documentation_Diagnostic_doc=[].verified.txt
@@ -1,7 +1,7 @@
 ﻿{
   Location: schema.avsc: (0,0)-(6,1),
   Message:
-The schema defined in the JSON is invalid: 'doc' property must be a string in schema: {
+The schema defined in the JSON is invalid: 'doc' property must be a string (found '[]') in schema: {
     "type": "fixed",
     "name": "Fixed",
     "namespace": "SchemaNamespace",
