+++ tests/AvroSourceGenerator.Tests/Snapshots/FixedTests/FixedTests.Verify_Name_Diagnostic_name=null.verified.txt
@@ -1,7 +1,7 @@
 ﻿{
   Location: schema.avsc: (0,0)-(5,1),
   Message:
-The schema defined in the JSON is invalid: 'name' property must be a string in schema: {
+The schema defined in the JSON is invalid: 'name' property must be a non-empty, non-whitespace string (found '') in schema: {
     "type": "fixed",
     "name": null,
     "namespace": "SchemaNamespace",
