--- tests/AvroSourceGenerator.Tests/Snapshots/FixedTests/FixedTests.Verify_Namespace_Diagnostic_namespace=--.verified.txt
@@ -1,19 +1,0 @@
-﻿{
-  Location: schema.avsc: (0,0)-(5,1),
-  Message:
-The schema defined in the JSON is invalid: 'namespace' property cannot be whitespace in schema: {
-    "type": "fixed",
-    "name": "Fixed",
-    "namespace": "",
-    "size": 16
-},
-  Severity: Error,
-  Descriptor: {
-    Id: AVROSG0002,
-    Title: Invalid Schema,
-    MessageFormat: The schema defined in the JSON is invalid: {0},
-    Category: Compiler,
-    DefaultSeverity: Error,
-    IsEnabledByDefault: true
-  }
-}
