+++ tests/AvroSourceGenerator.Tests/Snapshots/FixedTests/FixedTests.Verify_Namespace_Diagnostic_namespace=[].verified.txt
@@ -1,7 +1,7 @@
 ﻿{
   Location: schema.avsc: (0,0)-(5,1),
   Message:
-The schema defined in the JSON is invalid: 'namespace' property must be a string in schema: {
+The schema defined in the JSON is invalid: 'namespace' property must be a string (found '[]') in schema: {
     "type": "fixed",
     "name": "Fixed",
     "namespace": [],
