+++ tests/AvroSourceGenerator.Tests/Snapshots/FixedTests/FixedTests.Verify_Namespace_namespace=--.verified.txt
@@ -1,0 +1,20 @@
+﻿// <auto-generated/>
+#pragma warning disable CS8618, CS8633, CS8714, CS8775, CS8981
+    [global::System.CodeDom.Compiler.GeneratedCode("AvroSourceGenerator", "1.0.0.0")]
+    public partial class Fixed : global::Avro.Specific.SpecificFixed
+    {
+        public Fixed() : base(16) { }
+    
+        public uint Size { get => (uint)Value.Length; }
+    
+        public override global::Avro.Schema Schema { get => Fixed._SCHEMA; }
+        public static readonly global::Avro.Schema _SCHEMA = global::Avro.Schema.Parse("""
+        {
+          "type": "fixed",
+          "name": "Fixed",
+          "namespace": "",
+          "size": 16
+        }
+        """);
+    }
+
