+++ tests/AvroSourceGenerator.Tests/Snapshots/RecordTests/RecordTests.Verify_Namespace_namespace=--.verified.txt
@@ -1,0 +1,37 @@
+﻿// <auto-generated/>
+#pragma warning disable CS8618, CS8633, CS8714, CS8775, CS8981
+#nullable enable
+    [global::System.CodeDom.Compiler.GeneratedCode("AvroSourceGenerator", "1.0.0.0")]
+    public partial record Record : global::Avro.Specific.ISpecificRecord
+    {
+    
+        public global::Avro.Schema Schema { get => Record._SCHEMA; }
+        public static readonly global::Avro.Schema _SCHEMA = global::Avro.Schema.Parse("""
+        {
+          "type": "record",
+          "name": "Record",
+          "namespace": "",
+          "fields": []
+        }
+        """);
+    
+        public object? Get(int fieldPos)
+        {
+            switch (fieldPos)
+            {
+                default: throw new global::Avro.AvroRuntimeException($"Bad index {fieldPos} in Get()");
+            }
+        }
+        
+        public void Put(int fieldPos, object? fieldValue)
+        {
+            switch (fieldPos)
+            {
+                default:
+                    throw new global::Avro.AvroRuntimeException($"Bad index {fieldPos} in Put()");
+            }
+        }
+    }
+
+#nullable restore
+#pragma warning restore CS8618, CS8633, CS8714, CS8775
